Add checks for the referential integrity

+++ src/main/java/ch/fhnw/wodss/webapplication/components/contract/ContractService.java
@@ -1,5 +1,7 @@
 package ch.fhnw.wodss.webapplication.components.contract;
 
+import ch.fhnw.wodss.webapplication.components.employee.EmployeeDto;
+import ch.fhnw.wodss.webapplication.components.employee.EmployeeRepository;
 import ch.fhnw.wodss.webapplication.configuration.AuthenticatedEmployee;
 import ch.fhnw.wodss.webapplication.exceptions.EntityNotFoundException;
 import ch.fhnw.wodss.webapplication.exceptions.InternalException;
@@ -14,11 +16,19 @@ public class ContractService {
 
     private final ContractRepository contractRepository;
 
-    public ContractService(ContractRepository contractRepository) {
+    private final EmployeeRepository employeeRepository;
+
+    public ContractService(ContractRepository contractRepository, EmployeeRepository employeeRepository) {
         this.contractRepository = contractRepository;
+        this.employeeRepository = employeeRepository;
     }
 
     public ContractDto createContract(ContractDto contract, AuthenticatedEmployee authenticatedEmployee) {
+        Optional<EmployeeDto> selectedEmployee = employeeRepository.getEmployeeById(contract.getEmployeeId());
+        if (selectedEmployee.isEmpty()) {
+            throw new EntityNotFoundException("employee", contract.getEmployeeId());
+        }
+
         Optional<ContractDto> createdContract = contractRepository.saveContract(contract);
         if (createdContract.isEmpty()) {
             throw new InternalException("Unable to create the contract");
@@ -41,15 +51,30 @@ public class ContractService {
     }
 
     public ContractDto updateContract(Long id, ContractDto contract, AuthenticatedEmployee authenticatedEmployee) {
+        Optional<ContractDto> selectedContract = contractRepository.getContractById(id);
+        if (selectedContract.isEmpty()) {
+            throw new EntityNotFoundException("contract", id);
+        }
+
+        Optional<EmployeeDto> selectedEmployee = employeeRepository.getEmployeeById(contract.getEmployeeId());
+        if (selectedEmployee.isEmpty()) {
+            throw new EntityNotFoundException("employee", contract.getEmployeeId());
+        }
+
         Optional<ContractDto> updatedContract = contractRepository.updateContract(id, contract);
         if (updatedContract.isEmpty()) {
-            throw new EntityNotFoundException("contract", id);
+            throw new InternalException("Unable to update the contract");
         }
 
         return updatedContract.get();
     }
 
     public void deleteContract(Long id, AuthenticatedEmployee authenticatedEmployee) {
+        Optional<ContractDto> selectedContract = contractRepository.getContractById(id);
+        if (selectedContract.isEmpty()) {
+            throw new EntityNotFoundException("contract", id);
+        }
+
         contractRepository.deleteContract(id);
     }
 }
